Remove unnecessary link to unassigned isssues
Every issue should have a milestone, and this is currently enforced.
Most issues do not have assignees, but this is a byproduct of the devs'
workflow. The more accurate way to see issues where we *want* and
assignee but don't have one is by looking at the "help wanted" label,
which is already linked in this README.